+++ fuzz/fuzz_targets/terminal_capability.rs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-#![no_main]
-
-use libfuzzer_sys::fuzz_target;
-
-fuzz_target!(|input: &[u8]| {
-    suzumushi::terminal_capabilities::fuzz_parse(input);
-});